+++ src/main/java/com/daniil1380/restservice/dto/Answer.java
@@ -4,10 +4,10 @@ public class Answer {
 
     private String answer;
 
-    private Integer id;
+    private Long id;
 
 
-    public Answer(String answer, Integer id) {
+    public Answer(String answer, Long id) {
         this.answer = answer;
         this.id = id;
     }
@@ -23,11 +23,11 @@ public class Answer {
         this.answer = answer;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
